refactor: add tui package
