+++ apps/cli/src/components/NextStepsPreview.ts
@@ -1,6 +1,5 @@
+import { Container, Panel } from "@repo/tui";
 import { Ansi, Box } from "effect-boxes";
-import { Container } from "../lib/Layout.js";
-import { Panel } from "./Panel.js";
 
 const sectionTitle = (title: string) =>
   Box.text(title).pipe(Box.annotate(Ansi.combine(Ansi.bold, Ansi.cyan)));
