+++ packages/tui/package.json
@@ -1,0 +1,28 @@
+{
+  "name": "@repo/tui",
+  "version": "0.0.0",
+  "private": true,
+  "type": "module",
+  "exports": {
+    ".": "./src/index.ts"
+  },
+  "main": "./src/index.ts",
+  "types": "./src/index.ts",
+  "scripts": {
+    "type-check": "tsc --noEmit",
+    "scratch": "bun run ./scratchpad.ts",
+    "test": "vitest run",
+    "clean": "git clean -xdf .cache .turbo dist node_modules tsconfig.tsbuildinfo"
+  },
+  "dependencies": {
+    "@repo/domain": "workspace:*",
+    "effect": "4.0.0-beta.65",
+    "effect-boxes": "^0.14.0"
+  },
+  "devDependencies": {
+    "@effect/vitest": "4.0.0-beta.65",
+    "@types/node": "^25.7.0",
+    "@repo/config-typescript": "workspace:*",
+    "vitest": "^4.1.6"
+  }
+}
