+++ packages/tui/src/components/TextArea.ts
@@ -3,8 +3,8 @@ import { Prompt } from "effect/unstable/cli";
 import { Ansi, Box, Cmd } from "effect-boxes";
 import { KeyBinding, whenBinding } from "../lib/KeyBinding.js";
 import * as Viewport from "../lib/Viewport.js";
-import { Hint } from "./Hint.js";
-import { Panel, PromptChrome } from "./Panel.js";
+import { Hint } from "./atom/Hint.js";
+import { Panel, PromptChrome } from "./atom/Panel.js";
 
 const Action = Data.taggedEnum<Prompt.ActionDefinition>();
