+++ packages/tui/src/components/atom/Panel.ts
